detectTransferRestriction modified, removed all requires and using boolean comparisions

+++ contracts/IERC1404.sol
@@ -8,12 +8,12 @@ abstract contract IERC1404 {
     /// @param value Amount of tokens being transferred
     /// @return Code by which to reference message for rejection reasoning
     /// @dev Overwrite with your custom transfer restriction logic
-    function  detectTransferRestriction (address from, address to, uint256 value) public virtual view returns (uint8);
+    function  detectTransferRestriction (address from, address to, uint256 value) public virtual view returns (uint256);
 
     /// @notice Returns a human-readable message for a given restriction code
     /// @param restrictionCode Identifier for looking up a message
     /// @return Text showing the restriction's reasoning
     /// @dev Overwrite with your custom message and restrictionCode handling
-    function  messageForTransferRestriction  (uint8 restrictionCode) public virtual view returns (string memory);
+    function  messageForTransferRestriction  (uint256 restrictionCode) public virtual view returns (string memory);
 }
